// src/components/ChatComponent/src/types/types.ts
export interface IUser extends Partial<User> {
  id: string;
  name: string | null;
  userJID?: string | null;
  token?: string;
  refreshToken?: string;
}

export interface IMessage {
  id: string; // message ID (aka timestamp in microseconds)
  user: IUser;
  date: Date | string; // date converted from id / timestamp (e.g. "2024-02-18T03:24:33.102Z")
  body?: string; // message body
  roomJid: string; // room id
  key?: string; // workaround to solve a problem of messages uniqueness - additional, local timestamp to solve when XMPP server sends duplicate timestamps (TO DO: depricate / review)
  coinsInMessage?: string | number; // store only - message coins counter
  numberOfReplies?: number[] | number; // store only - array of replies in a thread (if applicable) - includes messages IDs so that client app can display relevant message previews for the thread
  isSystemMessage?: string;
  isMediafile?: string;
  locationPreview?: string;
  mimetype?: string;
  location?: string;
  pending?: boolean;
  timestamp?: number;
  showInChannel?: string;
  activeMessage?: boolean;
  isReply?: boolean | string;
  isDeleted?: boolean;
  mainMessage?: string;
  reply?: IReply[];
  fileName?: string;
  translations?: TranslationObject;
  langSource?: string;
  size?: string;
}

export interface IReply extends IMessage {}

export interface IRoom {
  id?: string;
  name: string;
  jid: string;
  title: string;
  usersCnt: number;
  messages: IMessage[];
  isLoading: boolean;
  roomBg: string;

  lastMessage?: string;
  lastRoomMessage?: RoomLastMessage;
  icon?: string;
  composing?: boolean;
  composingList?: string[];
  lastViewedTimestamp?: number;
  unreadMessages?: number;
  noMessages?: boolean;
  role?: string;

  roomMembers?: RoomMember[];
}

export interface RoomMember {
  ban_status: string;
  jid: string;
  last_active: number;
  name: string;
  role: string;
}

export interface RoomLastMessage {
  name: string;
  body: string;
}

export interface UserType extends IMessage {
  id: any;
  user: any;
  timestamp: any;
  text: any;
}

export interface ConfigUser {
  email: string;
  password: string;
}

export interface User {
  walletAddress: string;

  description?: string;
  token: string;
  refreshToken: string;

  defaultWallet: {
    walletAddress: string;
  };
  _id: string;
  firstName: string;
  lastName: string;
  email?: string;
  username?: string;
  profileImage?: string;
  emails?: [
    {
      loginType: string;
      email: string;
      verified: boolean;
      _id: string;
    },
  ];
  appId: string;
  xmppPassword: string;

  homeScreen?: string;
  registrationChannelType?: string;
  updatedAt?: string;
  authMethod?: string;
  resetPasswordExpires?: string;
  resetPasswordToken?: string;
  xmppUsername?: string;
  roles?: string[];
  tags?: string[];
  __v?: number;

  isProfileOpen?: boolean;
  isAssetsOpen?: boolean;
  isAgreeWithTerms?: boolean;
  isSuperAdmin?: any;
}

export interface XmppState {
  client: any;
  loading: boolean;
}

export interface IConfig {
  disableHeader?: boolean;
  disableMedia?: boolean;
  colors?: {primary: string; secondary: string};
  googleLogin?: {
    enabled: boolean;
    firebaseConfig: {
      apiKey: string;
      authDomain: string;
      projectId: string;
      storageBucket: string;
      messagingSenderId: string;
      appId: string;
    };
  };
  jwtLogin?: {
    token: string;
    enabled: boolean;
    handleBadlogin?: React.ReactElement;
  };
  userLogin?: {
    enabled: boolean;
    user: User | null;
  };
  customLogin?: {
    enabled: boolean;
    loginFunction: any; //() => Promise<User>
  };
  baseUrl?: string;
  xmppSettings?: xmppSettingsInterface;
  disableRooms?: boolean;
  defaultLogin?: boolean;
  disableInteractions?: boolean;
  chatHeaderBurgerMenu?: boolean;
  forceSetRoom?: boolean;
  roomListStyles?: ViewStyle;
  chatListStyles?: {
    background?: string;
    searchbackground?: string;
  };
  chatRoomStyles?: ViewStyle;
  setRoomJidInPath?: boolean;
  disableRoomMenu?: boolean;
  defaultRooms?: string[] | ConfigRoom[];
  refreshTokens?: {enabled: boolean; refreshFunction?: any};
  backgroundChat?: {
    color?: string;
    image?:
      | ImageSourcePropType
      | React.FC<React.SVGProps<SVGSVGElement>>
      | React.FC<SvgProps>
      | React.JSX.Element;
  };
  bubleMessage?: {
    backgroundMessageUser?: string;
    backgroundMessage?: string;
    colorUser?: string;
    color?: string;
    borderRadius?: number;
  };
  headerLogo?:
    | ImageSourcePropType
    | React.FC<React.SVGProps<SVGSVGElement>>
    | React.FC<SvgProps>
    | React.JSX.Element;
  emptyChatLogo?:
    | ImageSourcePropType
    | React.FC<React.SVGProps<SVGSVGElement>>
    | React.FC<SvgProps>
    | React.JSX.Element;
  headerMenu?: () => void;
  headerChatMenu?: (() => void) | boolean;
  customRooms?: {
    rooms: PartialRoomWithMandatoryKeys[];
    disableGetRooms?: boolean;
    singleRoom: boolean;
  };
  enableTranslates?: string;
  disableRoomConfig?: boolean;
  disableProfilesInteractions?: boolean;
  disableUserCount?: boolean;
  clearStoreBeforeInit?: boolean;
  disableRoomMenuModal?: boolean
}

export type PartialRoomWithMandatoryKeys = Partial<IRoom> &
  Pick<IRoom, 'jid' | 'title'>;

export interface xmppSettingsInterface {
  devServer: string;
  host: string;
  conference?: string;
}

interface ConfigRoom {
  jid: string;
  pinned: boolean;
  _id?: string;
}